% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensoDashUI.R
\name{SensoDashUI}
\alias{SensoDashUI}
\title{Launch SensoDash application}
\usage{
SensoDashUI()
}
\description{
A User Interface for the SensoDash package. The UI is a Shiny application for consumer data segmentation.
}
\examples{
library(SensoDash)

SensoDashUI()

}
